Support argument for output file path
This adds support for writing the output directly to a TXF file instead
of to stdout. The default behavior is still to print to stdout, however
if the script is invoked with two arguments then it will treat the
second argument as an output file.

This will only write to the output file if the file does not already
exist. This is to guard against overwriting existing tax documents by
accident.

Test: ./mssb_1099b_to_txf.py 1099.pdf > out1.txf
Test: ./mssb_1099b_to_txf.py 1099.pdf out2.txf
Test: diff out1.txf out2.txf